feat: simplify admin UI - dead 204 branch, unused fields, shared hook and time formatters (admin-ui-1,2,3,5)
