+++ admin/src/api.ts
@@ -35,10 +35,6 @@ async function request<T>(
     throw new Error(data.error || data.detail || `HTTP ${resp.status}`);
   }
 
-  if (resp.status === 204) {
-    return undefined as T;
-  }
-
   return resp.json() as Promise<T>;
 }
