+++ admin/src/format.ts
@@ -1,0 +1,9 @@
+export function formatTime(iso: string): string {
+  const d = new Date(iso);
+  const pad = (n: number) => String(n).padStart(2, "0");
+  return `${pad(d.getHours())}:${pad(d.getMinutes())}`;
+}
+
+export function formatDateTime(iso: string): string {
+  return `${formatTime(iso)} · ${new Date(iso).toLocaleDateString()}`;
+}
